Move `executeCommandLine` to `src/cli/execute.ts`
